fix(web): share pre-app shell on setup and join
Same logo column, card, and legal footer. Setup logo stays
unlinked so / cannot bounce a gated user.
