+++ apps/web/src/app/(flows)/layout.tsx
@@ -1,43 +1,9 @@
-import Link from "next/link";
-import { useTranslations } from "next-intl";
-
-import { SokosumiLogo, ThemedLogo } from "@/components/masumi-logos";
-import { ClientMessageBoundary } from "@/i18n/client-message-boundary";
-import { AUTH_MESSAGE_PATHS } from "@/i18n/message-namespaces";
+import { PreAppShell } from "@/components/pre-app-shell";
 
 export default function FlowsLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  return (
-    <ClientMessageBoundary paths={AUTH_MESSAGE_PATHS}>
-      <div className="flex h-svh gap-6 p-6">
-        <div className="flex h-full flex-1 flex-col gap-6">
-          <Link href="/">
-            <ThemedLogo LogoComponent={SokosumiLogo} priority />
-          </Link>
-          <div className="mx-auto flex w-full max-w-lg flex-1 items-center justify-center">
-            {children}
-          </div>
-          <FlowsLayoutFooter />
-        </div>
-      </div>
-    </ClientMessageBoundary>
-  );
-}
-
-function FlowsLayoutFooter() {
-  const t = useTranslations("Auth.Footer");
-
-  return (
-    <div className="flex items-center justify-center gap-4">
-      <Link href="/privacy-policy" className="text-sm hover:text-gray-300">
-        {t("privacyPolicy")}
-      </Link>
-      <Link href="/terms-of-service" className="text-sm hover:text-gray-300">
-        {t("termsOfServices")}
-      </Link>
-    </div>
-  );
+  return <PreAppShell>{children}</PreAppShell>;
 }
